+++ packages/abi/lib.esm/abi-coder.js
@@ -20,7 +20,6 @@ const paramTypeBytes = new RegExp(/^bytes([0-9]*)$/);
 const paramTypeNumber = new RegExp(/^(u?int)([0-9]*)$/);
 export class AbiCoder {
     constructor(coerceFunc) {
-        logger.checkNew(new.target, AbiCoder);
         defineReadOnly(this, "coerceFunc", coerceFunc || null);
     }
     _getCoder(param) {