+++ packages/providers/lib.esm/etherscan-provider.js
@@ -158,7 +158,6 @@ function checkError(method, error, transaction) {
 }
 export class EtherscanProvider extends BaseProvider {
     constructor(network, apiKey) {
-        logger.checkNew(new.target, EtherscanProvider);
         super(network);
         defineReadOnly(this, "baseUrl", this.getBaseUrl());
         defineReadOnly(this, "apiKey", apiKey || defaultApiKey);
@@ -175,9 +174,11 @@ export class EtherscanProvider extends BaseProvider {
                 return "https:/\/api-kovan.etherscan.io";
             case "goerli":
                 return "https:/\/api-goerli.etherscan.io";
+            case "optimism":
+                return "https:/\/api-optimistic.etherscan.io";
             default:
         }
-        return logger.throwArgumentError("unsupported network", "network", name);
+        return logger.throwArgumentError("unsupported network", "network", this.network.name);
     }
     getUrl(module, params) {
         const query = Object.keys(params).reduce((accum, key) => {