+++ packages/providers/src.ts/base-provider.ts
@@ -757,8 +757,6 @@ export class BaseProvider extends Provider implements EnsProvider {
      */
 
     constructor(network: Networkish | Promise<Network>) {
-        logger.checkNew(new.target, Provider);
-
         super();
 
         // Events being listened to